% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdlvm.R
\name{bdlvm}
\alias{bdlvm}
\alias{bdlvm_model_setup}
\title{\verb{bdlvm_*} functions for streamlined model generation/fitting}
\usage{
bdlvm_model_setup(formula, data, M)
}
\arguments{
\item{formula}{\code{brmsformula} specifying the "core" model.}

\item{data}{If there is no exogenous variables, an integer suffice to indicate you want a dataframe with the amount of rows as part of the output. Otherwise, a dataframe with the values of exogenous variables is expected and its number of rows will be used for the template data.}

\item{M}{The number of items to be added on each latent variable. Currently a single number i.e. all the latent variables will get the same amount of items.}
}
\description{
These functions have been designed with the following scenario in mind:
\itemize{
\item Someone has a specific SEM in mind they wish to test
\item They may want to generate data from a variety of parameter values
\item They may want to fit to that data using a variety of model specifications
}

Therefore, it makes sense to break up each of those steps into different functions, so that e.g. a model needs to be specified only once, while generation and fitting are each specified separately. In the general case, then one can have a total of \code{S*D*F} sets of results, where \code{S} is the number of SEMs, \code{D} is the number of datasets and \code{F} is the number of fitting specifications.

See details for the workflow in terms of specific functions. Note these package is currently designed to work solely on top of \code{brms} and \code{SBC}.
}
\details{
The specific workflow is as follows:
\enumerate{
\item Provide a description of the SEM "core" (i.e. exogenous and latent variables only) via a \code{brms} formula. Any variables tagged with the \code{mi()} function are assumed to be latent variables.
\item Feed the "base formula" into \code{bdlvm_model_setup} to produce a model object that will have a "full formula" and template dataset that has added variables corresponding to the items for each latent variable.
\item Feed the model object to \code{bdlvm_make_data} together with the number of desired datasets to simulate the data (this is a SLOW STEP as it uses \code{brms::posterior_predict} iteratively to populate the entire graph).
\item Feed the dataset and model objects to \code{bdlvm_compute} in order to fit the model to each dataset via a modified version of \code{SBC::compute_SBC}. The same generative priors will be used for fitting by default, but see the arguments for ways to specify different priors. EXPERIMENTAL: the backend object SBC requires will be generated at this point in order to guarantee compilation of the latest model specification. YES I LEARNED THIS IN A VERY PAINFUL MANNER.
}
}
\section{Functions}{
\itemize{
\item \code{bdlvm_model_setup()}: Make a MODEL OBJECT out of a BASE FORMULA

}}
\examples{
suppressMessages(library(brms))
bdlvm_model_setup(bf(y1|mi()~1), data = 50, M = 3)
}
