% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdlvm.R
\name{bdlvm_add_items}
\alias{bdlvm_add_items}
\title{Make a MODEL OBJECT out of a BASE FORMULA}
\usage{
bdlvm_add_items(formula, data, M, ...)
}
\description{
A base formula specifies only the exogenous and latent variables in a model. By passing through this function, a model object is generated, which includes a formula and template dataset with items corresponding to each latent variable added.
}
\examples{
suppressMessages(library(brms))
bdlvm_add_items(bf(y1|mi()~1) +
  bf(y21|mi()~mi(y1)) + bf(y22|mi()~mi(y1)) +
  bf(y3|mi()~mi(y21)+mi(y22)), data = 5, M = 3)
}
