% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdlvm.R
\name{bdlvm_add_prior}
\alias{bdlvm_add_prior}
\title{Make GENERATIVE PRIORS out of a base model object (formula+data)}
\usage{
bdlvm_add_prior(
  model_obj,
  mu = 0,
  sigma = 1,
  betas = 1,
  lf = 1,
  noise = 1,
  item_mu = 0,
  explain = FALSE,
  prior_only = FALSE,
  ...
)
}
\value{
object with edited formula and data + priors unless priors_only = TRUE
}
\description{
look at arguments for defaults (constant values on all parameters)
}
\examples{
suppressMessages(library(brms))
bdlvm_add_items(bf(y1|mi()~1) +
  bf(y21|mi()~mi(y1)) + bf(y22|mi()~mi(y1)) +
  bf(y3|mi()~mi(y21)+mi(y22)), data = 5, M = 3) -> sem_formula

  bdlvm_add_prior(sem_formula, prior_only = TRUE)
}
