% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdlvm.R
\name{bdlvm_brms_pregen}
\alias{bdlvm_brms_pregen}
\title{Prepare arguments required for a SEM \code{SBC} generator via \code{brms}}
\usage{
bdlvm_brms_pregen(
  formula,
  data = 1,
  M = 1,
  mu = 0,
  sigma = 1,
  betas = 1,
  lf = 1,
  noise = 1,
  item_mu = 0,
  add_items = TRUE,
  mode = "predict",
  explain = FALSE,
  ...
)
}
\arguments{
\item{formula}{SEM formula specifying exogenous and latent variables. Any variable using the \code{mi()} syntax is assumed to be a latent variable and will be assigned receive priors and item variables according to the arguments that follow.}

\item{data}{A template dataframe that will be passed to \code{brm}. The number of rows determines the number of observations that will be generated. Only values for exogenous variables are used, all latent variables will be overwritten by the generated values. If no exogenous variables are present, a single integer can be passed to indicate the number of observations to generate.}

\item{M}{The amount of items per latent variable. It is currently assumed to be identical for all.}

\item{mu, sigma}{Intercepts and variances for the latent variables. Can be fixed values or strings describing valid prior distributions to be used by \code{brms} for parameter draws. Must be of length 1 or equal to the number of latent variables in the formula.}

\item{betas}{Coefficients for regressions between latent variables. Currently only supporting linear  predictors on the location parameters. Fixed values or strings, as above. Length not checked but should be equal to amount of relationships in the model, see \code{explain} argument for more.}

\item{lf, noise, item_mu}{Loading factors, variances and intercepts for the items. Fixed values or strings, as above. Must be of length 1, M or equal to the total number of items in the model.}

\item{add_items}{\code{TRUE} by default. Will cause item terms and priors to be added to the output model formula, dataframe and priors.}

\item{mode}{Default value of \code{"predict"}. Generates output with the expectation that \code{brms} will be used to draw parameter values and then generate data via \code{posterior_predict}. Can be changed to \code{"fit"}, which will instead fit the full model to generate values via \code{mi()} imputation.}

\item{explain}{Will print text explaining the order in which priors will be assigned to model parameters.}

\item{...}{Currently unused.}
}
\value{
A list with \code{formula}, \code{data} and \code{prior} components, ready to be passed into \code{SBC_generator_brms}. Be aware that priors for exogenous variables are left unspecified, so the user must finish setting them up in order to produce sensible results.
}
\description{
Prepare arguments required for a SEM \code{SBC} generator via \code{brms}
}
\examples{
library(brms)
bdlvm_brms_pregen(
 bf(y1|mi()~1) +
   bf(y21|mi()~mi(y1)) + bf(y22|mi()~mi(y1)) +
   bf(y3|mi()~mi(y21)+mi(y22)),
 M = 3, data = mtcars[1:5,1:2],
 sigma = c(10, 2, 2, 2), noise = 0.01,
 explain = TRUE) -> .bdlvm_pregen_example

 .bdlvm_pregen_example$formula
 head(.bdlvm_pregen_example$prior, 10)
}
