% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdlvm.R
\name{bdlvm_compute}
\alias{bdlvm_compute}
\title{Title}
\usage{
bdlvm_compute(
  model_obj,
  data_obj,
  ...,
  debug_mode = FALSE,
  priors = NULL,
  var_constraint = FALSE,
  cache_mode = "none",
  cache_location = NULL
)
}
\arguments{
\item{model_obj}{An object made with \code{bdlvm_model_setup}}

\item{data_obj}{An object made with \code{bdlvm_generate_datasets}}

\item{...}{Arguments passed to \code{brms::brm}}

\item{priors}{Named list passed to \code{make_fit_prior}, see details for accepted names}

\item{cache_mode}{Disabling SBC cache by default}

\item{cache_location}{Disabling SBC cache by default}
}
\value{
Just a regular \code{SBC_results}-class object.
}
\description{
Bottom text
}
\details{
Names in the list passed via the \code{priors} argument can refer to SEM-specific parameters by being one of
\code{item_intercept}, \code{item_sigma}, \code{item_loading}, \code{lv_intercept}, \code{lv_sigma}, \code{lv_to_lv_coef},
\code{lv_d_intercept}, \code{lv_to_lv_d_coef}, \code{lv_other}, \code{lv_d_other}

One can pass a \code{sem_global} argument to assign a prior on all non-intercept
parameters that were not specifically named. Additionally, the name \code{non_sem}
can be used to refer to any other parameters that are not part of the basic SEM
structure; this include exogenous variables and \code{I()} transformations of latent variables.
}
\examples{
suppressMessages({
  library(brms)
  library(lunafuns)
})

# general model setup
base_formula <- bf(y1|mi()~1) +
  bf(y21|mi()~mi(y1)) + bf(y22|mi()~mi(y1)) +
  bf(y3|mi()~mi(y21)+mi(y22))

model_object <- bdlvm_model_setup(base_formula, data = 50, M = 3)

# data generation setup
dataset_object <- bdlvm_make_data(model_object, n_sims = 2)

# fit (with a potentially different prior)
fits_object <- bdlvm_compute(model_object, dataset_object,
                             priors = list(
                               global = "normal(0,5)"
                             ))
}
