% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdlvm.R
\name{bdlvm_compute_SBC}
\alias{bdlvm_compute_SBC}
\title{Title}
\usage{
bdlvm_compute_SBC(backend_obj, data_obj, varmethod = "varmi", ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
# simple example of workflow using bdlvm helpers
suppressMessages({
  library(brms)
  library(lunafuns)
})

# general model setup
base_formula <- bf(y1|mi()~1) +
  bf(y21|mi()~mi(y1)) + bf(y22|mi()~mi(y1)) +
  bf(y3|mi()~mi(y21)+mi(y22))

model_object <- bdlvm_add_items(base_formula, data = 5, M = 5)

# backend setup
backend_object <- bdlvm_make_backend(model_object)

# data generation setup
generator_arguments <- bdlvm_add_prior(model_object)
dataset_object <- bdlvm_generate_sem(generator_arguments, n_sims = 2)

# fit mode
fits <- bdlvm_compute_SBC(backend_object, dataset_object,
                          global_prior = "normal(0,3)")
}
