% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdlvm.R
\name{bdlvm_make_backend}
\alias{bdlvm_make_backend}
\title{Make a BACKEND out of a MODEL OBJECT}
\usage{
bdlvm_make_backend(model_obj, ..., backend = "cmdstanr", chains = 4)
}
\arguments{
\item{model_obj}{Object created via bdlvm_add_items}

\item{...}{Extra arguments passed to the brms backend}

\item{backend}{Stan backend choice, defaults to \code{"cmdstanr"}}

\item{chains}{Defaults to 4}
}
\value{
List with an SBC backend object, latent variable names and item names
}
\description{
Make a BACKEND out of a MODEL OBJECT
}
\examples{
# no
}
