% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_arg_maker.R
\name{create_arg_maker}
\alias{create_arg_maker}
\title{Make a function to generate argument tables}
\usage{
create_arg_maker(...)
}
\arguments{
\item{...}{Argument names for the desired function, possibly including default values}
}
\value{
A function taking only the arguments passed as \code{...}, which generates a tibble with the argument names as columns.
}
\description{
Make a function to generate argument tables
}
\examples{
create_arg_maker(N, M, mu = 0, sigma = 1)
}
