% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{depmat_ordlist}
\alias{depmat_ordlist}
\title{Create a list of variables ordered according to their dependency matrix}
\usage{
depmat_ordlist(dep_mat)
}
\arguments{
\item{dep_mat}{A dependency matrix made with \code{fomrlst_depmat}}
}
\value{
A list where each element contains a vector of variable names that depend on variables in the previous level. The first element holds variables with no dependencies.
}
\description{
Create a list of variables ordered according to their dependency matrix
}
\examples{
{
  testf <- list(x11 ~ y1, x12 ~ y1, x13 ~ y1,
                x21 ~ y2, x22 ~ y2, x23 ~ y2,
                x31 ~ y3, x32 ~ y3, x33 ~ y3,
                 y3 ~ y1, y2 ~ y1 + z21 + z22)
  depmat_ordlist(formlst_depmat(testf))
}
}
