% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{formlst_depmat}
\alias{formlst_depmat}
\title{Create a dependency matrix from a list of formulas}
\usage{
formlst_depmat(formula_list)
}
\arguments{
\item{formula_list}{A list containing \code{formula} class objects}
}
\value{
A square matrix, with names and size equal to the unique variables identified in the formulas. For a given variable's row, values of \code{1} indicate the variables that depend it. For a given variable's column, values of \code{1} indicate the variables it depends on.
}
\description{
Create a dependency matrix from a list of formulas
}
\examples{
{
  testf <- list(x11 ~ y1, x12 ~ y1, x13 ~ y1,
                x21 ~ y2, x22 ~ y2, x23 ~ y2,
                x31 ~ y3, x32 ~ y3, x33 ~ y3,
                 y3 ~ y1, y2 ~ y1 + z21 + z22)
  formlst_depmat(testf)
}
}
