% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_lhs_mi}
\alias{get_lhs_mi}
\title{Indicate which formulas contain \verb{|mi()} on the LHS}
\usage{
get_lhs_mi(formula_list)
}
\arguments{
\item{formula_list}{A list of formulas}
}
\value{
A logical vector that returns \code{TRUE} where \verb{|mi()} is found in the formula's LHS.
}
\description{
Indicate which formulas contain \verb{|mi()} on the LHS
}
\examples{
get_lhs_mi(list(y ~ x, z | mi() ~ y, w ~ mi(y)))
}
