% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_lhs_vars}
\alias{get_lhs_vars}
\title{Get the names of all LHS variables}
\usage{
get_lhs_vars(formula_list)
}
\arguments{
\item{formula_list}{A list of formulas}
}
\value{
A character vector with the names of unique variables on each LHS, in order of appearance.
}
\description{
Get the names of all LHS variables
}
\examples{
get_lhs_vars(list(y~x, z~y, w+z~x))
}
