% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{qist}
\alias{qist}
\title{Get a quoted list}
\usage{
qist(...)
}
\arguments{
\item{...}{List contents, can be named}
}
\value{
An expression to create a list with the given contents
}
\description{
Designed for use with \code{targets::tar_map()}, see https://github.com/ropensci/tarchetypes/discussions/105
}
\examples{
 expr1 <- qist(arg1 = 123, arg2 = letters[1:3], "done")
 expr2 <- quote(list(arg1 = 123, arg2 = letters[1:3], "done"))
 identical(expr1,expr2)
}
