% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_helpers.R
\name{sbch_brms_backend}
\alias{sbch_brms_backend}
\title{Construct \code{brms} backend for SBC}
\usage{
sbch_brms_backend(..., template_data, ..dropFILLER = TRUE)
}
\arguments{
\item{...}{Arguments passed to \code{brm}}

\item{template_data}{Sample of data \code{brms} will work with

If a function is passed as \code{formula} or \code{prior} arguments via \code{...}, this
helper will send the result of applying the function on \code{template_data} to
the backend instead.}
}
\description{
Construct \code{brms} backend for SBC
}
