% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_helpers.R
\name{sbch_cmdstan_backend}
\alias{sbch_cmdstan_backend}
\title{Construct \code{cmdstanr} backend for SBC}
\usage{
sbch_cmdstan_backend(model_path, exe_path)
}
\arguments{
\item{model_path}{Stan model file location}

\item{exe_path}{Compiled model executable location}
}
\value{
Object of class \code{SBC_backend_cmdstan_sample} with an added \code{model_name} attribute
}
\description{
Construct \code{cmdstanr} backend for SBC
}
