% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_helpers.R
\name{sbch_generate}
\alias{sbch_generate}
\title{Adds more flexibility to the SBC generator}
\usage{
sbch_generate(generator, ..., n_sims = NULL, n_reps = NULL, mvbrms = TRUE)
}
\arguments{
\item{generator}{SBC generator function}

\item{n_sims}{Number of datasets to simulate (usual behaviour)}

\item{n_reps}{Number of times to repeat a single simulated dataset (for other MCMC assessment purposes)}

\item{mvbrms}{\code{FALSE} will use the default \code{SBC::generate_datasets} function. By default, we use our modified version that can run multivariate \code{brms} models.}
}
\value{
Datasets in the format SBC expects
}
\description{
Adds more flexibility to the SBC generator
}
