% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_helpers.R
\name{sbch_generate_mvdatasets}
\alias{sbch_generate_mvdatasets}
\title{Modified version of \code{SBC::generate_datasets} that runs with multivariate \code{brms} models}
\usage{
sbch_generate_mvdatasets(generator, n_sims, n_datasets = NULL)
}
\arguments{
\item{generator}{A generator built with \code{SBC_generator_brms}}

\item{n_sims}{Number of simulated datasets to produce}
}
\value{
Object of class \code{SBC_datasets} (contains generated data and true parameter values)
}
\description{
Modified version of \code{SBC::generate_datasets} that runs with multivariate \code{brms} models
}
\examples{
print("nope")
}
