% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_helpers.R
\name{sbch_generate_semdata}
\alias{sbch_generate_semdata}
\title{Modified \code{SBC::generate_datasets} to make SEM data from \code{brms} models}
\usage{
sbch_generate_semdata(generator, n_sims, n_datasets = NULL)
}
\arguments{
\item{generator}{hmm}

\item{n_sims}{Number of simulated datasets to produce}
}
\value{
Object of class \code{SBC_datasets} (contains generated data and true parameter values)
}
\description{
Modified \code{SBC::generate_datasets} to make SEM data from \code{brms} models
}
\examples{
print("nope")
}
