% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_helpers.R
\name{sbch_generator}
\alias{sbch_generator}
\title{Construct SBC generator from a function and arguments list}
\usage{
sbch_generator(gen_fun, arg_list, ...)
}
\arguments{
\item{gen_fun}{The generator function}

\item{arg_list}{List of named arguments that will be passed to the generator}

\item{...}{Other parameters passed to the generator}
}
\value{
An object of class \code{SBC_generator_function}
}
\description{
Construct SBC generator from a function and arguments list
}
