% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc_helpers.R
\name{sbch_run}
\alias{sbch_run}
\title{\code{compute_SBC} wrapper}
\usage{
sbch_run(datasets, backend)
}
\arguments{
\item{datasets}{an object of class \code{SBC_datasets}}

\item{backend}{model sampler made by some \verb{SBC_backend_*}function}
}
\description{
\code{compute_SBC} wrapper
}
