% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_helpers.R
\name{sh_compile_path}
\alias{sh_compile_path}
\title{Compile Stan model to desired location}
\usage{
sh_compile_path(source, where = ".")
}
\arguments{
\item{source}{Path to a Stan model file}

\item{where}{Folder where the compiled executable will be placed}
}
\value{
The relative path of the compiled executable
}
\description{
Compile Stan model to desired location
}
