% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{which_order}
\alias{which_order}
\title{Get indexes matching \code{y} with each value of \code{x}}
\usage{
which_order(x, y)
}
\arguments{
\item{x, y}{Some vectors}
}
\value{
A vector of indices of elements in \code{y} that match \code{x}, in the order given by \code{x}
}
\description{
Get indexes matching \code{y} with each value of \code{x}
}
\examples{
set.seed(1)
which_order(c("b", "d"), sample(letters[1:4], 10, replace = TRUE))
}
